Update new hook
